+++ test/example/wrapper/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    get: function(req, res, next) { // jshint ignore:line
+        res.send('GET index ' + req.foo);
+    }
+};